+++ src/commands/dev/panel.ts
@@ -14,13 +14,8 @@ const Panel: Command[] = [{
   example: 'client',
   code: `
     $onlyIf[$checkContains[$botOwnerID[true;,];$authorID];]
-
-    $if[$guildID==;
-      $!httpRequest[https://api.ipify.org/?format=json;GET]
-      $let[ip;$httpResult[ip]]
-    ;
-      $let[ip;REDACTED]
-    ]
+    $defer
+    $if[$guildID==;$!httpRequest[https://api.ipify.org/?format=json;GET] $let[ip;$httpResult[ip]];$let[ip;REDACTED]]
 
     $color[$getGlobalVar[color]]
     $title[Developer Panel:]
@@ -29,13 +24,13 @@ const Panel: Command[] = [{
     $addField[Process Info;\n- <:Clock:$if[$environment==main;1263809772478529576;1271543787637833759]> Uptime: <t:$round[$math[$math[$getTimestamp-$uptime]/1000];0]:R>\n- <:CPU:$if[$environment==main;1263809787661910050;1271543827395772517]> CPU: \`$cpuUsage[false]% {$cpuCores}\`\n- <:Server:$if[$environment==main;1263809924295819285;1271544093838938235]> Ram: \`$ram[true;false]\`;true]
     $addField[;;true]
     $addField[Client Info;**Client Stats:**\n- <:Roles:$if[$environment==main;1263809915529728072;1271544076491292736]> Guilds: \`$guildCount\`\n- <:members:$if[$environment==main;1263809852145270926;1271543939576627322]> Users: \`$userCount\`\n- <:shard:$if[$environment==main;1263809940569456650;1271544119080259666]> Shards: \`$shardsOnline/$shardCount\`\n- <:Plus:$if[$environment==main;1263809898878210149;1271544047579693148]> Commands: \`$commandCount\`;true]
-    $addField[<:Spacer:$if[$environment==main;1324808760207736842;1275843251349356675]>;<:Spacer:$if[$environment==main;1324808760207736842;1275843251349356675]>\n- <:Akira_Active_Dev:$if[$environment==main;1263809801817821278;1271546151543373906]> CMDs since startup: \`$getGlobalVar[startCommands]\`\n- <:ping:$if[$environment==main;1263809890468499508;1271544029053452297]> DB Latency: \`$dbPing\`\n- <:ping:$if[$environment==main;1263809890468499508;1271544029053452297]> WS Latency: \`$pingms\`\n- <:ping:$if[$environment==main;1263809890468499508;1271544029053452297]> RT Latency: \`$roundtrip\`;true]
+    $addField[<:Spacer:$if[$environment==main;1324808760207736842;1275843251349356675]>;<:Spacer:$if[$environment==main;1324808760207736842;1275843251349356675]>\n- <:ping:$if[$environment==main;1263809890468499508;1271544029053452297]> DB Latency: \`$dbPing\`\n- <:ping:$if[$environment==main;1263809890468499508;1271544029053452297]> WS Latency: \`$pingms\`\n- <:ping:$if[$environment==main;1263809890468499508;1271544029053452297]> RT Latency: \`$roundtrip\`;true]
     $addField[;;true]
 
     $addActionRow
-    $addButton[panelPrevious~$authorID~0;;Secondary;◀️;true]
+    $addButton[panelButton~$authorID~0;;Secondary;◀️;true]
     $addButton[panelGuild~$authorID;Guild;Secondary;;$if[$guildID==;true;false]]
-    $addButton[panelNext~$authorID~0;;Secondary;▶️;false]
+    $addButton[panelButton~$authorID~1;;Secondary;▶️;false]
     $addButton[close~$authorID~message~$messageID;Close;Danger;✖️]
   `
 }];