+++ src/commands/events/onInteractionCreate.ts
@@ -11,7 +11,23 @@ const interactionCreate: Command[] = [
     example: 'N/A',
     version: '1.0.0',
     code: `
-    $c[$logger[Info;Ran $customID!]]
+    $if[$isSlashCommand;
+      $setGlobalVar[sessionSlashUses;$math[$getGlobalVar[sessionSlashUses]+1]]
+      $setGlobalVar[monthSlashUses;$math[$getGlobalVar[monthSlashUses]+1]]
+      $setGlobalVar[weekSlashUses;$math[$getGlobalVar[weekSlashUses]+1]]
+      $setGlobalVar[daySlashUses;$math[$getGlobalVar[daySlashUses]+1]]
+      $setGlobalVar[totalSlashUses;$math[$getGlobalVar[totalSlashUses]+1]]
+      $if[$guildID!=;
+        $setGuildVar[totalSlashUses;$math[$getGuildVar[totalSlashUses;$customEncrypt[encrypt;$guildID]]+1];$customEncrypt[encrypt;$guildID]]
+      ]
+    ]
+    $if[$isButton;
+      $setGlobalVar[totalButtonUses;$math[$getGlobalVar[totalButtonUses]+1]]
+      $setGlobalVar[monthButtonUses;$math[$getGlobalVar[monthButtonUses]+1]]
+      $setGlobalVar[weekButtonUses;$math[$getGlobalVar[weekButtonUses]+1]]
+      $setGlobalVar[dayButtonUses;$math[$getGlobalVar[dayButtonUses]+1]]
+      $setGlobalVar[sessionButtonUses;$math[$getGlobalVar[sessionButtonUses]+1]]
+    ]
     `
   }
 ];