+++ src/commands/events/onReady.ts
@@ -8,7 +8,10 @@ const onReady: Event[] = [{
   version: '1.0',
   code: `
     $logger[Info;$username[$clientID] | Running with "$commandCount" commands]
-    $setGlobalVar[startCommands;0]
+    $setGlobalVar[sessionMessageUses;0]
+    $setGlobalVar[sessionSlashUses;0]
+    $setGlobalVar[sessionButtonUses;0]
+    $setGlobalVar[sessionErrors;0]
   `
 }];